+++ src/game_api/entity.hpp
@@ -59,7 +59,7 @@ struct EntityHooksInfo
     std::vector<HookWithId<void(Entity*)>> on_destroy;
     std::vector<HookWithId<void(Entity*, Entity*)>> on_kill;
     std::vector<HookWithId<bool(Entity*)>> on_player_instagib;
-    std::vector<HookWithId<bool(Entity*, Entity*, int8_t, float, float, uint8_t, uint8_t)>> on_damage;
+    std::vector<HookWithId<bool(Entity*, Entity*, int8_t, float, float, uint16_t, uint8_t)>> on_damage;
     std::vector<HookWithId<bool(Movable*)>> pre_statemachine;
     std::vector<HookWithId<void(Movable*)>> post_statemachine;
     std::vector<HookWithId<void(Container*, Movable*)>> on_open;
@@ -295,7 +295,7 @@ class Entity
     void set_on_destroy(std::uint32_t reserved_callback_id, std::function<void(Entity*)> on_destroy);
     void set_on_kill(std::uint32_t reserved_callback_id, std::function<void(Entity*, Entity*)> on_kill);
     void set_on_player_instagib(std::uint32_t reserved_callback_id, std::function<bool(Entity*)> on_instagib);
-    void set_on_damage(std::uint32_t reserved_callback_id, std::function<bool(Entity*, Entity*, int8_t, float, float, uint8_t, uint8_t)> on_damage);
+    void set_on_damage(std::uint32_t reserved_callback_id, std::function<bool(Entity*, Entity*, int8_t, float, float, uint16_t, uint8_t)> on_damage);
     void set_pre_collision1(std::uint32_t reserved_callback_id, std::function<bool(Entity*, Entity*)> pre_collision1);
     void set_pre_collision2(std::uint32_t reserved_callback_id, std::function<bool(Entity*, Entity*)> pre_collision2);
